Moved utilities for main step testing
