--- test/helpers/init.js
@@ -1,7 +1,0 @@
-const runAction = require('./run-action.js')
-const steps = ['pre', 'main']
-
-module.exports = () => {
-  runAction(steps)
-  return new Promise(resolve => setTimeout(resolve, 1500))
-}
