+++ test/main/file-exists.js
